Moved `field` model to its own folder
