+++ dplib/models/__init__.py
@@ -1,8 +1,9 @@
 from .contributor import Contributor
 from .dialect import Dialect
+from .field import Constraints, Field
 from .license import License
 from .package import Package
 from .profile import Profile
 from .resource import Resource
-from .schema import Constraints, Field, ForeignKey, ForeignKeyReference, Schema
+from .schema import ForeignKey, ForeignKeyReference, Schema
 from .source import Source
