+++ dplib/models/schema/__init__.py
@@ -1,5 +1,3 @@
-from .constraints import Constraints
-from .field import Field
 from .foreignKey import ForeignKey
 from .foreignKeyReference import ForeignKeyReference
 from .schema import Schema
